requirements.txt
allure-pytest==2.5.2
allure-python-commons==2.5.2
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
configparser==3.5.0
crayons==0.1.2
future==0.16.0
idna==2.7
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pytest==3.9.2
pytest-ordering==0.6
requests==2.20.0
selene==1.0.0a13
selenium==2.53.1
six==1.11.0
urllib3==1.24
webdriver-manager==1.7
